+++ pages/dashboard.js
@@ -58,7 +58,12 @@ class Dashboard extends Component {
 
   render () {
     if (this.state.loading) {
-      return <LoadingState />
+      return (
+        <div>
+          <header className='header--internal--green header--page' style={{ paddingBottom: '8rem' }} />
+          <LoadingState />
+        </div>
+      )
     }
     const { authenticatedUser } = this.props
     const { loggedIn, account } = authenticatedUser
@@ -189,7 +194,7 @@ class Dashboard extends Component {
         </section>
         <section>
           <div className='row widget-container'>
-            <DashboardSidebar teams={teams} osmesaData={osmesaData} />
+            <DashboardSidebar teams={teams} osmesaData={osmesaData} loggedIn={loggedIn} />
             <div className='widget-75'>
               <DashboardBadges badges={badges} name={name} />
             </div>