+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## Unreleased
+
+## [v1.7.0] - 2019-11-28
 ### Added
 - Add country of origin column to campaign tables
 
@@ -20,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Replace `edit_sum` on User pages with the updated `edit_count` from OSMesa
 - Fixes bug in tables when defaults for statistics are not provided by the API
 - Remove redundant tooltips from campaign tables
+
 ## [v1.6.0] - 2019-10-28
 ### Changed
 - Move all table header labels to common json file with descriptions.
@@ -249,7 +251,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The first release
 
-[Unreleased]: https://github.com/developmentseed/scoreboard/compare/v1.6.0...HEAD
+[Unreleased]: https://github.com/developmentseed/scoreboard/compare/v1.7.0...HEAD
+[v1.7.0]: https://github.com/developmentseed/scoreboard/compare/v1.6.0...v1.7.0
 [v1.6.0]: https://github.com/developmentseed/scoreboard/compare/v1.5.0...v1.6.0
 [v1.5.0]: https://github.com/developmentseed/scoreboard/compare/v1.4.0...v1.5.0
 [v1.4.0]: https://github.com/developmentseed/scoreboard/compare/v1.3.3...v1.4.0